# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skills API",
        "description": "API endpoints for CRUD operations on skills",
        "version": "1.0.0"
    },
    "paths": {
        "/api/skills": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get a list of skills",
                "operationId": "6acb38b223d6471cc32eab45ea57be4b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query for filtering skills by title and slug using full-text search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new skill",
                "operationId": "4725cd560d3046ae33bd690208e21d41",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Skill Title"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "skill-slug"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill Created"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/skills/{id}": {
            "get": {
                "summary": "Get a single skill by ID",
                "operationId": "81b9cbc27bbdd8b5a6591975bf75441a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the skill",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a skill by ID",
                "operationId": "6c932a7a6f40be346ab93e4dd28f14f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the skill",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Skill Title"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "skill-slug"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill Updated"
                    },
                    "404": {
                        "description": "Skill not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a skill by ID",
                "operationId": "7401a82a4b0de61e80c67901e89d62dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the skill",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill Deleted"
                    },
                    "404": {
                        "description": "Skill not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login as an admin user",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "403": {
                        "description": "Email or password is incorrect or user is not an admin"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the currently authenticated user",
                "operationId": "67d5147477c548bfbc3f982833bbc7bd",
                "responses": {
                    "204": {
                        "description": "Logout successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get the currently authenticated user's details",
                "operationId": "2130b7cce3aebae5b2f0aecd8f3cf0d3",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}